Added files
One file is for the final purposes which is our Final_Code.  The other file is for the purpose of our demonstration.  Our Demo_Code file will have the timers for limited amounts of time in order to show how the final product will work in a shortened amount of time.